more file existence checks

+++ mylar/PostProcessor.py
@@ -185,8 +185,12 @@ class PostProcessor(object):
                     shutil.move(path_to_move, os.path.join(mylar.DUPLICATE_DUMP, file_to_move))
                 except (OSError, IOError):
                     logger.warn('[DUPLICATE-CLEANUP] Failed to move ' + path_to_move + ' ... to ... ' + os.path.join(mylar.DUPLICATE_DUMP, file_to_move) + ' - so deleted the file ')
-                    os.remove(path_to_move)
-                    dontcheckdupe=1
+                    if os.path.isfile(path_to_move):
+                        os.remove(path_to_move)
+                        dontcheckdupe=1
+                    else:
+                        dontcheckdupe=1
+                        logger.warn('[DUPLICATE-CLEANUP] Failed to move ' + path_to_move + ' ... to ... ' + os.path.join(mylar.DUPLICATE_DUMP, file_to_move) + ' - file was already deleted, apparently ')
                     return True
 
                 logger.warn('[DUPLICATE-CLEANUP] Successfully moved ' + path_to_move + ' ... to ... ' + os.path.join(mylar.DUPLICATE_DUMP, file_to_move))
